Move the source-neutral parts of serialpps into a pps package
serialpps mixed two things: reading a serial port's modem-control pins,
and the parts that do not care where a pulse comes from, namely the edge
types, the adaptive polling loop that finds a pulse on a pin whose level
can only be read, and the generator that labels an edge with a UTC second
from the receiver's time messages. A GPIO pulse source needs the second
group but not the first.

Create gps/app/pps holding Edge, CandidateEdge, Sample, Generator,
PollStats and Poll. Poll now takes a PulseReader, whose InPulse method
answers the one question the loop asks; serialpps adapts a StateReader and
its Wiring to it. The generator's two delay keys move to
pps.GeneratorConfig, which serialpps.Config embeds; go-toml inlines an
untagged embedded struct, so the [sample.serial.pps] keys are unchanged.
check.Validate learns the same rule so its messages keep naming the keys
as they appear in the file. No behaviour changes.
