+++ gps/app/pps/bubble_windows_test.go
@@ -1,4 +1,4 @@
-package serialpps
+package pps
 
 import "testing"
