+++ gps/app/pps/readtime.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package serialpps
+package pps
 
 import "time"
