+++ gps/app/pps/readtime_windows.go
@@ -1,4 +1,4 @@
-package serialpps
+package pps
 
 import (
 	"time"
