+++ gps/app/pps/sleep_linux.go
@@ -1,4 +1,4 @@
-package serialpps
+package pps
 
 import "time"
