+++ gps/app/serialpps/config.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/jclark/satpulse/gps/app/gpsio"
+	"github.com/jclark/satpulse/gps/app/pps"
 	"github.com/jclark/satpulse/gps/lib/check"
 )
 
@@ -13,11 +14,7 @@ import (
 // UTC-labelled receiver messages. Durations are expressed in seconds in
 // TOML.
 type Config struct {
-	// DelayUncertainty is the allowed measurement uncertainty when an
-	// inferred post-pulse message delay is slightly negative.
-	DelayUncertainty float64 `toml:"delayUncertainty" check:">=0,<1" comment:"Uncertainty in measured pulse-to-message delay (s)"`
-	// MaxDelay is the maximum accepted inferred post-pulse message delay.
-	MaxDelay float64 `toml:"maxDelay" check:">0,<1" comment:"Maximum post-pulse message delay (s)"`
+	pps.GeneratorConfig
 	// Method selects how edges are detected; unspecified means automatic.
 	Method gpsio.PPSMethod `toml:"method" comment:"PPS edge detection method: poll, wait, or kernel; omit for automatic selection"`
 	// MaxWakeupLatency optionally limits latency added when a CPU wakes from
@@ -31,10 +28,7 @@ type Config struct {
 
 // DefaultConfig returns the default serial PPS sampling configuration.
 func DefaultConfig() Config {
-	return Config{
-		DelayUncertainty: 0.005,
-		MaxDelay:         0.8,
-	}
+	return Config{GeneratorConfig: pps.DefaultGeneratorConfig()}
 }
 
 // Validate checks that the delay interval is valid and narrower than one
